changed c3 to use int instead of long long int

+++ 3/c.c
@@ -5,7 +5,7 @@ int main(int argc, char *argv[]){
   printf("%d\n", prime);
 }
 int bprime(long long int n){
-  long long int i = 2;
+  int i = 2;
   while(i < n){
     if(n%i==0){
       n /= i;
